獲得用のエフェクトを追加。 フラワーサークル獲得時のアニメ演出とエフェクトの生成処理を追加。 デバッグ開始。
