+++ Assets/Materials/eff_par_1.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 1eb614596030841cbbdb034483a7b79c
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
