+++ Assets/Materials/eff_wind_1.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 85ff0e8ed75f542a2a6009d16b39c939
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
