+++ Assets/New Material.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 8179a267cc8f89148aee26020b2095b0
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
