+++ Assets/Prefab/Eff_Heal_1.prefab.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 214bf6b1f924e43108a92f13c4db31a2
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
