+++ Assets/Scripts/FlowerCircle.cs
@@ -1,0 +1,63 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+using DG.Tweening;
+
+public class FlowerCircle : MonoBehaviour
+{
+    public int point;
+
+    //private GameObject[] flowerObjs;
+
+    [SerializeField]
+    private GameObject effectPrefab;
+
+    //[SerializeField]
+    //private BoxCollider boxCollider;
+
+    void Start()
+    {
+        transform.DORotate(new Vector3(0, 360, 0), 5f, RotateMode.FastBeyond360).SetEase(Ease.Linear).SetLoops(-1, LoopType.Restart);
+        //flowerObjs = GetAllChildren();
+    }
+
+    //private GameObject[] GetAllChildren() {
+    //    GameObject[] objs = new GameObject[transform.childCount];
+    //    for (int i = 0; i < transform.childCount; i++) {
+    //        objs[i] = transform.GetChild(i).gameObject;
+    //    }
+    //    return objs;
+    //}
+
+    private void OnTriggerEnter(Collider other) {
+
+        transform.SetParent(other.transform);
+
+        StartCoroutine(PlayGetEffect());        
+    }
+
+    private IEnumerator PlayGetEffect() {
+        //boxCollider.enabled = false;
+        
+
+        Sequence sequence = DOTween.Sequence();
+        sequence.Append(transform.DOScale(Vector3.zero, 1.0f));
+        sequence.Join(transform.DOLocalMove(new Vector3(0, 0, 0), 1.0f));
+
+        //for (int i = 0; i < flowerObjs.Length; i++) {
+        //    Sequence sequence = DOTween.Sequence();
+        //    //sequence.Append(flowerObjs[i].transform.DOMoveY(-10.0f, 0.25f)).SetRelative();
+        //    sequence.Append(flowerObjs[i].transform.DOScale(Vector3.zero, 1.0f));   // .OnComplete(() => { Destroy(flowerObjs[i]); })
+        //    sequence.Join(flowerObjs[i].transform.DOLocalMove(new Vector3(0, 0, 0), 1.0f));
+        //}
+
+        yield return new WaitForSeconds(1.0f);
+
+        // エフェクト
+        GameObject effect = Instantiate(effectPrefab, transform.position, Quaternion.identity);
+        effect.transform.position = new Vector3(effect.transform.position.x, effect.transform.position.y -1.5f, effect.transform.position.z);
+        Destroy(effect, 1.0f);
+
+        Destroy(gameObject, 1.0f);
+    }
+}
